larc-migration/scripts/requirements.txt
# Requirements pour le Sync Daemon
supabase>=2.0.0
python-dotenv>=1.0.0
